// src/interfaces/main.interface.ts
export type IMain = {
    id: number,
    name: string
}

export type IModelMain = {
    id: number;
    name: string;
    created_at:number;
    updated_at:number;
    is_active:boolean;
    is_delete:boolean
}

export type IEsMain = {
    id: number;
    name: string;
    created_at: number;
    updated_at: number;
    is_active: boolean;
}